requirements.txt
numpy
pandas
matplotlib
tqdm
scikit-learn
streamlit
Pillow
pyarrow
protobuf